Modify RayTracer.m for stats
The main changes from RayTracer.m concern the paths for reading the CAD files, as the folder /stats contains one folder per scenario. In each scenario folder, the CADop.mat and *.xml files are contained. At each run of  collectStats.m, a new subfolder /$timestamp is created, containing the default /Input and /Output structure, with the exception of the aforementioned CAD files.